Added stats.py and pstat.py files from WWW
These files appear to be available only via these URLs:

http://www.nmr.mgh.harvard.edu/Neural_Systems_Group/gary/python/stats.py

http://www.nmr.mgh.harvard.edu/Neural_Systems_Group/gary/python/pstat.py

According to http://www.nmr.mgh.harvard.edu/Neural_Systems_Group/gary/python.html,
these files are the precursor to the scipy.stats package.
corr_functions.py depends on the functionality of these older scripts;
it is not compatible with scipy.stats.

It would be preferable to modify corr_functions.py to use the
scipy.stats package, but that is beyond the scope of the current
exercise (resurrecting the PAPER South Africa 64 input correlator).