Eliminate global variables in rx module
The rx module contained global variables.  For some weird reason, this
caused weird bugs wirh compiled on Ubuntu 12.04.1 LTS using GCC 4.6.3.
The symptom was that incrementing the global variable "packet_count"
would actually increment the global variable "sock", rendering it
impossible to send display data using "sock" as the socket file
descriptor.  This behavior was seen using gdb on cn_rx.py by setting a
watchpoint on "sock" and finding that the watch point was triggered
whenever packet_count was incremented.  Weird!  I don't know whether it
is a bug in GCC 4.6.3, a bug in the Linux dynamic linker/loader, or just
a natural consequence of having global variables in a shared library
(admittedly NOT a good idea anyway!).

This commit moves the global variables to be file static and function
static variables.  Things still compile fine (so the reduced scope is
not problematic) and the weird behavior described above is now gone.