Revert "Fix channel computation for GPU X engine"
This reverts commit bc135cf14fc73b48eb727f83130a64d717d9508e.
That commit hard-coded the ratio of F:X engines as 2:1.  Reverting that
commit means that the default ratio of F:X is now 1:1.  For instruments
that have an F:X ratio of N:1 (N>1), the recommendation is to scale
"n_ants_per_feng" in the configuration file by a factor of N.