Change how we set the source IP address for F->GPU
The source address for the F-Engine-to-GPU 10 GbE cores used to be a
single fixed IP address.  This worked OK for when all the destinations
IP addresses were on the same subnet as the single fixed IP address.
This is no longer the case with the new GPU 10 GbE interface address
assignments where 10 GbE interface ethN is given IP address 192.168.N.X,
where X is a host-specific value for all 10 GbE (and other) interfaces
on a given host.

We now compute our own F-Engine-to-GPU source IP address by changing the
last octect of the destination IP address to 254.  Note that this is
better than using a single, fixed source IP address for all
F-engine-to-GPU 10 GbE cores because it allows the destination addresses
to be in different subnets, but it still results in multiple interfaces
having the same source IP address so, as with using a single IP address
for all such interfaces, it is not really suitable for use with a switch
between the F Engines and the X boxes (it is only for direct connect).