Added more details to nfeng != nxeng error message
If the number of F engines specified in the borphserver/servers config
item differs from the number of X engines specified in the
correlator/gpu_ip config item, the config code ignores ALL gpu_ip values
and prints an error message about the mismatch.  The old error message
did not indicate the names of the items in the config file that were
mismatches.  The new error message now does.

Ignoring ALL gpu_ip values is probably not the best thing to do since
it can (and does in feng_init.py) lead to an indexing error.  Probably
best to error out if the number of F engines exceeds the number of X
engines.  If the number of X engines exceeds the number of F engines,
then print warning, but proceed.  Or maybe best to error out if any
mimatch at all?