Reuse data and flags buffers in cn_rx.py
Previously, each visibility spectrum and its associated flags array were
freshly allocated from the heap, filled in, and then passed to the
callback function which assumed ownership of (i.e. eventually freed)
them.  Since the most common use case does not require the callback
function to retain the spectrum and flag data, the allocation strategy
has been changed so that the buffers are allocated once and reused over
and over again.  Callback functions do NOT assume ownership of these
buffers.  If a callback function ever does wish to retain the data, it
must copy them to buffers that the callback function manages.