Avoid static array size for xids in collate_packet
collate_packet() has an array of integers used for counting packets from
each X engine.  This array was statically sized for a given correlator
configuration of (up to) 16 X engines.  This is obviously a problem for
larger correlators with more X engines, so this array is now dynamically
sized at runtime for the actual number of X engines as computed from
parameters in the conf file.