Add some profiling/benchmarking code
With a 256 input 1024 channel correlator the code now takes over 20
seconds to write out an integration.  This is a big problem for 10
second integrations!  To find out where the code is taking so long,
various portions of the code have been surrounded by clock_gettime calls
to measure the time spent in those areas.

The python code now has this performance code always enabled, but the C
code must be compiled with "PACKET_STATS" defined to have this code
included.