Disable unused/unknown "sdisp_callback"
It turns out that the sdisp_callback is causing the dump to take too
long.  The sdisp_callback re-formats and re-transmits the data for
reasons that are no longer known.  There are no known receivers of the
data being sent by sdisp_callback.  For these reasons, sdisp_callback
have been disabled.