Add support for hookup and antpos files
This adds support for arbitrary correlator hookups to make the system
antenna-to-input agnostic.  It also includes "real" antenna positions in
the MIRIAD datasets rather than dummy placeholder coordinates (A,A,A)
for antenna A.  See the comments in cn_rx.py for details about how the
antenna positions are stored (units, frame, etc.).

This relies on having the following two
configuration mapping files (from the "papercfg" repository) installed:

/etc/papercfg/psa128/ant_to_pos.yml - antenna to position
/etc/papercfg/psa128/fxin_to_antpol.yml - correlator input to antpol

On paper1, the papercfg repository is cloned into "/home/obs/papercfg".