Move relabel logic earlier in filewrite_callback
This moves the relabeling logic to the very beginning of
filewrite_callback.  This is needed because the relabeling need to
happen as early as possible in that function, *before* any of the
antenna or polarization info are used.

This also ensures that the baselines being written are in the proper
order (i.e. a1<=a2) (and cross-polarization XY vs YX).