Make uv files full stokes
Attempt to manually revert ca45293 to go back to
full-stokes files.
This change was made in parallel with the addition of
metadata to the uv files (stupidly). This commit has attempted
to separate out the polarization-specific changes, but see the previous
commit on this branch in case it turns out (it will) this hasn't
entirely worked